requirements.txt
backports-abc==0.5
bokeh==0.12.3
certifi==2016.9.26
click==6.6
Django==1.10.3
Flask==0.11.1
futures==3.0.5
image==1.5.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
nltk==3.2.1
numpy==1.11.2
oauthlib==2.0.1
Pillow==3.4.2
plotly==1.12.10
python-dateutil==2.6.0
pytz==2016.7
PyYAML==3.12
requests==2.12.2
requests-oauthlib==0.7.0
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.2
tweepy==3.5.0
Werkzeug==0.11.11
wordcloud==1.2.1
